Should be done with this round on the Electron Efficiency, moving aways from TResult
Former-commit-id: fb653429e7e74a29bbb43ca3ef1a5b6b0cd320e6

+++ PhysicsAnalysis/ElectronPhotonID/ElectronEfficiencyCorrection/Root/TElectronEfficiencyCorrectionTool.cxx
@@ -171,7 +171,7 @@ int Root::TElectronEfficiencyCorrectionTool::initialize() {
         if (m_seed == 0) {
             std::unique_ptr<TMD5> tmd=CxxUtils::make_unique<TMD5>();
             const char* tmd_as_string=tmd->FileChecksum(fname.get())->AsString();
-            m_seed = *reinterpret_cast<const int*>(tmd_as_string);
+            m_seed = *(reinterpret_cast<const unsigned long int*>(tmd_as_string));
             ATH_MSG_INFO("Seed (automatically) set to " << m_seed);
         }else {
             ATH_MSG_INFO("Seed set to " << m_seed);
